KN-43 added more logic and is able to show the image from the klayout on the user interface on refresh. Will work on dynamic load tomorrow
