+++ website/templates/meshGenerator.html
@@ -149,7 +149,7 @@
     </head>
     <body>  
         <div class="topbar">
-            <img src="/static/SiClarity-Logo.png" alt="SiClarity-Logo"/>
+            <img src="/static/images/SiClarity-Logo.png" alt="SiClarity-Logo"/>
             <h1>MELD</h1>
         </div>
         <div class="inner-body">
